changes class name and its references.
