+++ src/sportsradar/extract/primaryfeeds.py
@@ -8,7 +8,7 @@ load_dotenv('.../../.env')
 logger = logging_helpers.get_logger(__name__)
 
 
-class ScheduleFeeds():
+class PrimaryFeeds():
     """ This class is responsible for extracting game feeds from SportRader """
 
     def __init__(self, base_url):
@@ -19,7 +19,7 @@ class ScheduleFeeds():
         """
 
         self.base_url = base_url
-    
+
     def get_current_season_schedule(self, access_level, version, language_code, file_format, api_key):
         """
         Get the current season schedule
@@ -40,7 +40,6 @@ class ScheduleFeeds():
         logger.info('Data retrieved successfully')
         return result
 
-    
     def get_current_week_schedule(self, access_level, version, language_code, file_format, api_key):
         """
         Get the current week schedule
@@ -57,7 +56,7 @@ class ScheduleFeeds():
             raise ValueError("API key not found in environment variables")
         datastore = DataStore(SportsRadarFetcher())
         result = datastore.fetch_data(
-            url = f"{self.base_url}/{access_level}/{version}/{language_code}/games/current_week/schedule.{file_format}?api_key={api_key}"
+            url=f"{self.base_url}/{access_level}/{version}/{language_code}/games/current_week/schedule.{file_format}?api_key={api_key}"
         )
         logger.info('Data retrieved successfully')
         return result
@@ -78,14 +77,14 @@ class ScheduleFeeds():
             logger.error("API key not found in environment variable.")
             raise ValueError("API key not found in environment variables")
         datastore = DataStore(SportsRadarFetcher())
-        result =datastore.fetch_data(
-            url = f"{self.base_url}/{access_level}/{version}/{language_code}/games/{year}/{season_type}/schedule.{file_format}?api_key={api_key}"
+        result = datastore.fetch_data(
+            url=f"{self.base_url}/{access_level}/{version}/{language_code}/games/{year}/{season_type}/schedule.{file_format}?api_key={api_key}"
 
         )
         return result
-    
 
-    def get_weekly_schedule(self, access_level, version, season_year, season_type, week_number, language_code, file_format, api_key):
+    def get_weekly_schedule(self, access_level, version, season_year, season_type, week_number, language_code,
+                            file_format, api_key):
         """
         Get Weekly Schedule
 
@@ -104,7 +103,7 @@ class ScheduleFeeds():
             logger.error("API key not found in enviroment variable.")
             raise ValueError("API key not found in environment variables")
         datastore = DataStore(SportsRadarFetcher())
-        result=datastore.fetch_data(
+        result = datastore.fetch_data(
             url=f"{self.base_url}/{access_level}/{version}/{language_code}/games/{season_year}/{season_type}/{week_number}/schedule.{file_format}?api_key={api_key}"
         )
